obj: bump block offset in memory block to 4 bytes
Up until recently the largest possible block offset inside of a run
was ~2500. This changed with the introduction of the flexible bitmap
and it allowed for creation of allocation classes with extremely
large "units_per_block" parameter. The new maximum is
max alloc size divided by the unit size. This new maximum exceeded
the uint16_t variable that stored this information in the runtime
state causing ASSERTs in debug builds and double allocation in
release ones. Instead of disallowing large units_per_block for certain
parameters, the runtime state variable has been changed to uint32_t
which is more than enough to store any possible units_per_block value.

+++ src/libpmemobj/container_seglists.c
@@ -50,7 +50,7 @@
 struct block_container_seglists {
 	struct block_container super;
 	struct memory_block m;
-	VECQ(, uint16_t) blocks[SEGLIST_BLOCK_LISTS];
+	VECQ(, uint32_t) blocks[SEGLIST_BLOCK_LISTS];
 	uint64_t nonempty_lists;
 };
 
@@ -108,7 +108,7 @@ container_seglists_get_rm_block_bestfit(struct block_container *bc,
 	/* finds the list that serves the smallest applicable size */
 	i = util_lssb_index64(v);
 
-	uint16_t block_offset = VECQ_DEQUEUE(&c->blocks[i]);
+	uint32_t block_offset = VECQ_DEQUEUE(&c->blocks[i]);
 
 	if (VECQ_SIZE(&c->blocks[i]) == 0) /* marks the list as empty */
 		c->nonempty_lists &= ~(1ULL << (i));
